% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zaggregate.R
\name{zaggregate}
\alias{zaggregate}
\title{Group and aggregate. Zuud!}
\usage{
zaggregate(.data, groups, ...)
}
\arguments{
\item{.data}{A data frame.}

\item{groups}{A vector of columns to group by.}

\item{...}{Aggregate functions.}
}
\value{
A tibble
}
\description{
Wrapper around dtplyr to simplify call to speedier aggregation.
}
