% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zbind.R
\name{zbind}
\alias{zbind}
\title{Bind data. Zuud!}
\usage{
zbind(x)
}
\arguments{
\item{x}{A dataframe..}
}
\value{
A tibble
}
\description{
Speedily binds a list of rows or dataframes into one.
}
\examples{

zbind(list(mtcars, mtcars, mtcars))

x <- list(a = 1, b = 1:5)
zbind(list(x, x, x))
}
