% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zread.R
\name{zread}
\alias{zread}
\title{Combine all dataframes in a folder. Zuud!}
\usage{
zread(dir, pattern = NULL, parallel = TRUE)
}
\arguments{
\item{dir}{Directory to search.}

\item{pattern}{Optional regex pattern to filter.}

\item{parallel}{If \code{TRUE} (and available on the machine), \code{zread()} will attempt to
use parallel processing. This may incur a small overhead cost that outweighs
the benefits on small datasets.}
}
\value{
A tibble.
}
\description{
\code{zread()} speedily reads and binds all files in a directory matching a regex pattern.
}
\details{
Accepted filetypes are csvs, tsvs, and feather files.
}
